move guardian management tests to GuardianStorage.spec.ts
